refacto(infra): move zammad infro code
