+++ bin/src/index.ts
@@ -1,6 +1,6 @@
 //
 
-import config from "@~/common.config";
+import config from "@~/app.core/config";
 import { LogLevels, consola } from "consola";
 import { showRoutes } from "hono/dev";
 import app from "../../src/api";
