+++ packages/~/app/core/src/config/index.ts
@@ -1,0 +1,9 @@
+//
+
+import env from "./env";
+
+//
+
+const ASSETS_PATH = `/assets/${env.VERSION}` as const;
+
+export default { ...env, ASSETS_PATH };
