+++ packages/~/app/core/src/error/index.ts
@@ -1,0 +1,8 @@
+//
+
+import ModernError from "modern-errors";
+
+//
+
+export const AuthError = ModernError.subclass("AuthError");
+export const HTTPError = ModernError.subclass("HTTPError");
