+++ packages/~/config/typescript/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@~/config.typescript",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "files": [],
+  "dependencies": {
+    "@tsconfig/bun": "1.0.4",
+    "@tsconfig/strictest": "2.0.3",
+    "bun-types": "1.0.28"
+  }
+}
