+++ packages/~/infra/zammad/api/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@~/zammad.api",
+  "version": "1.0.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": {
+      "default": "./src/index.ts"
+    }
+  },
+  "dependencies": {
+    "@hono/zod-validator": "0.1.11",
+    "@~/app.core": "workspace:*",
+    "@~/zammad.lib": "workspace:*",
+    "hono": "4.0.2",
+    "zod": "3.22.4"
+  },
+  "devDependencies": {
+    "@~/config.typescript": "workspace:*"
+  }
+}
