+++ packages/~/infra/zammad/api/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "@tsconfig/bun/tsconfig.json",
+  "compilerOptions": {
+    "composite": true,
+    "declaration": true,
+    "emitDeclarationOnly": true,
+    "lib": ["ESNext", "DOM"],
+    "noEmit": false,
+    "outDir": "./node_modules/.cache/tsc"
+  }
+}
