+++ packages/~/infra/zammad/lib/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@~/zammad.lib",
+  "version": "1.0.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": {
+      "default": "./src/index.ts"
+    },
+    "./*": {
+      "default": "./src/*.ts"
+    }
+  },
+  "dependencies": {
+    "@~/app.core": "workspace:*",
+    "consola": "3.2.3",
+    "zod": "3.22.4"
+  },
+  "devDependencies": {
+    "@~/config.typescript": "workspace:*"
+  }
+}
