+++ packages/~/infra/zammad/lib/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "compilerOptions": {
+    "outDir": "./node_modules/.cache/tsc"
+  },
+  "extends": "@~/config.typescript/bun.json"
+}
