--- src/common/errors.ts
@@ -1,7 +1,0 @@
-//
-
-import errorCustomClass from "error-custom-class";
-
-//
-
-export const HTTPError = errorCustomClass("HTTPError");
