+++ src/moderations/:id/Duplicate_Warning.tsx
@@ -4,8 +4,8 @@ import { OpenInZammad, SearchInZammad } from ":common/zammad";
 import { schema } from ":database:moncomptepro";
 import type { moncomptepro_pg_Context } from ":database:moncomptepro/middleware";
 import { app_hc } from ":hc";
-import { get_zammad_mail } from ":legacy/services/zammad_api";
 import { get_duplicate_moderations } from ":moderations/repository";
+import { get_zammad_mail } from "@~/zammad.lib";
 import to from "await-to-js";
 import { eq } from "drizzle-orm";
 import type { NodePgDatabase } from "drizzle-orm/node-postgres";
