+++ src/moderations/:id/email/Message.tsx
@@ -4,11 +4,9 @@ import { ASSETS_PATH } from ":assets/config";
 import { date_to_string } from ":common/date";
 import env from ":common/env";
 import { type Moderation } from ":database:moncomptepro";
-import {
-  GROUP_MONCOMPTEPRO_SENDER_ID,
-  type Article,
-} from ":legacy/services/zammad_api";
 import { quote } from ":ui/quote";
+import { GROUP_MONCOMPTEPRO_SENDER_ID } from "@~/zammad.lib/const";
+import type { Article } from "@~/zammad.lib/types";
 import { tv } from "tailwind-variants";
 
 //
